Got model loading working
Had to do super inefficient implementation because copy constructors are scary...
Will fix that up if it becomes a performance bottleneck

Also want to place asset management into a class that handles all of that.
User should only request from the manager to load a file and the manager should return a pointer to the object loaded.